Fix a potential dos venue by validating if path exists or is creatable for remotely saving data (yamls) Validate if paths we try to read are valid as well before attempting open()

+++ server/area_manager.py
@@ -202,6 +202,8 @@ class AreaManager:
 
     def load_music(self, path):
         try:
+            if not os.path.isfile(path):
+                raise AreaError(f'File path {path} is invalid!')
             with open(path, 'r', encoding='utf-8') as stream:
                 music_list = yaml.safe_load(stream)
 
@@ -232,6 +234,8 @@ class AreaManager:
 
         """
         try:
+            if not os.path.isfile(path):
+                raise
             with open(path, 'r') as chars:
                 data = yaml.safe_load(chars)
         except:
@@ -253,6 +257,8 @@ class AreaManager:
 
         """
         try:
+            if not os.is_path_exists_or_creatable(path):
+                raise ArgumentError(f'File path {path} cannot be created!')
             with open(path, 'w', encoding='utf-8') as stream:
                 yaml.dump(self.character_data, stream, default_flow_style=False)
         except:
